Fix trajectory and correct get_e_spec
This changes the definition of xcur, ycur, and zcur. Before they
were grid indicies now they are positions measured in km. That means
that their use in get_e_spec has to change. At the same time I also
updated it to use a spherical region of space instead of a cube.