Fix #1
The SWAP bins only need to be read once, so that has been moved to the
main program and the bins are now passed into get_e_spec as an argument.
Building the histogram with SWAPs logarithmic bins is done correctly on
the first pass rather than using an evenly spaced histogram as an
intermediary.